徴兵時の資源消費で、資源が65535を超えている場合、動作に不具合が起こる問題を修正（チェリーピック）

+++ SangokuKmy/Models/Commands/SoldierCommand.cs
@@ -113,7 +113,7 @@ namespace SangokuKmy.Models.Commands
             var needResources = add;
             foreach (var resource in resources)
             {
-              var newResource = (ushort)Math.Max(0, resource.Item.Resource - needResources);
+              var newResource = Math.Max(0, resource.Item.Resource - needResources);
               onSucceeds.Add(async () => await SpendResourceAsync(resource.Item, resource.Info, newResource));
 
               needResources -= resource.Item.Resource;
@@ -195,7 +195,7 @@ namespace SangokuKmy.Models.Commands
           {
             discountMoney += (int)(Math.Min(resource.Item.Resource, needResources) * moneyPerSoldier * (resource.Effect.Value / 100.0f));
 
-            var newResource = (ushort)Math.Max(0, resource.Item.Resource - needResources);
+            var newResource = Math.Max(0, resource.Item.Resource - needResources);
             onSucceeds.Add(async () => await SpendResourceAsync(resource.Item, resource.Info, newResource));
 
             needResources -= resource.Item.Resource;
@@ -217,7 +217,7 @@ namespace SangokuKmy.Models.Commands
             needResources = add;
             foreach (var resource in resources)
             {
-              var newResource = (ushort)Math.Max(0, resource.Item.Resource - needResources);
+              var newResource = Math.Max(0, resource.Item.Resource - needResources);
               onSucceeds.Add(async () => await SpendResourceAsync(resource.Item, resource.Info, newResource));
 
               needResources -= resource.Item.Resource;
@@ -267,7 +267,7 @@ namespace SangokuKmy.Models.Commands
       }
 
 
-      async Task SpendResourceAsync(CharacterItem item, CharacterItemInfo info, ushort newResource)
+      async Task SpendResourceAsync(CharacterItem item, CharacterItemInfo info, int newResource)
       {
         item.Resource = newResource;
         if (item.Resource <= 0)
